Added MPI utils functions for passing SFM datastructures. Preliminarily working in paralell

+++ parallel/src/types.hpp
@@ -7,7 +7,7 @@ namespace SFM {
 
   using KeyPoints = std::vector<cv::KeyPoint>;
   using Points = std::vector<cv::Point2f>;
-  using KeypointID = size_t;
+  using KeypointID = int;
   using LandmarkID = size_t;
 
   struct Features {
